Spring and rest (#5)
* created database for storing json

* changed files according to comments

* Added files service

* Revised File based database and Id service

* Small revision in Id service

* Added one test for better coverage

* Added spring boot classes

* Revision of files

* Revision of files

+++ src/main/java/pl/futurecollars/invoicing/db/file/FileBasedDatabase.java
@@ -79,6 +79,7 @@ public class FileBasedDatabase implements Database {
       String updatedInvoiceAsJson = jsonService.objectToString(invoiceToUpdate);
       allLines.add(updatedInvoiceAsJson);
       filesService.writeLinesToFile(invoicesPath, allLines);
+
     } catch (IOException exception) {
       throw new RuntimeException("Updating invoice failed for id: " + id, exception);
     }
@@ -94,12 +95,13 @@ public class FileBasedDatabase implements Database {
           .collect(Collectors.toList());
 
       filesService.writeLinesToFile(invoicesPath, updatedList);
+
     } catch (IOException exception) {
       throw new RuntimeException("Deleting invoice failed");
     }
   }
 
-  private boolean containsId(String line, long id) {
+  private boolean containsId(String line, int id) {
     return line.contains("\"id\":" + id + ",");
 
   }
